refactor: rename widgets/v2 → widgets/public
The staging directory was never intended to carry a version label.
Updates all import paths, test paths, and CLAUDE.md references.
