+++ src/bootstack/widgets/public/exceptions.py
