+++ src/bootstack/widgets/public/primitives/__init__.py
@@ -1,0 +1,3 @@
+from bootstack.widgets.public.primitives.button import Button
+
+__all__ = ["Button"]
