--- src/bootstack/widgets/v2/__init__.py
@@ -1,24 +1,0 @@
-from bootstack.widgets.v2.events import Event
-from bootstack.widgets.v2.exceptions import BootstackV2Error, UnknownEventError, ParentResolutionError
-from bootstack.widgets.v2.subscription import Subscription
-from bootstack.widgets.v2.container import PublicContainer
-from bootstack.widgets.v2.base import PublicWidgetBase
-from bootstack.widgets.v2.stacks import HStack, VStack
-from bootstack.widgets.v2.grid import Grid
-from bootstack.widgets.v2.app import App
-from bootstack.widgets.v2.primitives.button import Button
-
-__all__ = [
-    "App",
-    "Button",
-    "BootstackV2Error",
-    "Event",
-    "Grid",
-    "HStack",
-    "ParentResolutionError",
-    "PublicContainer",
-    "PublicWidgetBase",
-    "Subscription",
-    "UnknownEventError",
-    "VStack",
-]
