--- src/bootstack/widgets/v2/primitives/__init__.py
@@ -1,3 +1,0 @@
-from bootstack.widgets.v2.primitives.button import Button
-
-__all__ = ["Button"]
